move product-list component to the root package
