+++ sample-app/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { FormsModule} from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { ProductListComponent } from './product-list/product-list.component';
+import { ProductListComponent } from './components/product-list/product-list.component';
 import { ConvertToSpacesPipe } from './pipes/convert-to-spaces.pipe';
 import { StarComponent } from './star/star.component';
 
@@ -18,7 +18,8 @@ import { StarComponent } from './star/star.component';
   imports: [
     BrowserModule,
      FormsModule,
-     HttpClientModule
+     HttpClientModule,
+     
   ],
   providers: [],
   bootstrap: [AppComponent]
