+++ sample-app/src/app/components/product-list/product-list.component.scss
